+++ e2e/tests/accounts.test.ts
@@ -1,5 +1,8 @@
-import { utils, constants } from "ethers";
 
+import { utils, constants, Wallet } from "ethers";
+
+import { MSEC_PER_DAY, MSEC_PER_SEC } from '../../packages/common/src/utils/timestamp';
+import { CoreSDK } from "../../packages/core-sdk/src";
 import {
   initCoreSDKWithFundedWallet,
   ensureCreatedSeller,
@@ -17,7 +20,7 @@ const sellerWallet = seedWallet3; // be sure the seedWallet is not used by anoth
 
 describe("CoreSDK - accounts", () => {
   describe("dispute resolver", () => {
-    const escalationResponsePeriodInMS = 60_000_000_000;
+    const escalationResponsePeriodInMS = 90 * MSEC_PER_DAY - 1 * MSEC_PER_SEC;
     // TODO: use valid metadata uri
     const metadataUri = "ipfs://dispute-resolver-uri";
     const ethDisputeResolutionFee = {