+++ packages/common/src/configs.ts
@@ -1,4 +1,4 @@
-import { EnvironmentType, ProtocolConfig } from "./types";
+import { EnvironmentType, Lens, ProtocolConfig } from "./types";
 
 const chainIdToInfo = new Map<number, ProtocolConfig["nativeCoin"]>([
   [1234, { decimals: "18", name: "Ether", symbol: "ETH" }],
@@ -21,25 +21,21 @@ const chainIdToGraphTx = new Map<number, (txHash?: string) => string>([
 ]);
 
 // https://docs.lens.xyz/docs/deployed-contract-addresses
-const chainIdToLensContractAddresses = new Map<
-  number,
-  {
-    LENS_HUB_CONTRACT: string;
-    LENS_PERIPHERY_CONTRACT: string;
-  }
->([
+const chainIdToLensInfo = new Map<number, Lens>([
   [
     80001,
     {
       LENS_HUB_CONTRACT: "0x60Ae865ee4C725cd04353b5AAb364553f56ceF82",
-      LENS_PERIPHERY_CONTRACT: "0xD5037d72877808cdE7F669563e9389930AF404E8"
+      LENS_PERIPHERY_CONTRACT: "0xD5037d72877808cdE7F669563e9389930AF404E8",
+      apiLink: "https://api-mumbai.lens.dev/"
     }
   ],
   [
     137,
     {
       LENS_HUB_CONTRACT: "0xDb46d1Dc155634FbC732f92E853b10B288AD5a1d",
-      LENS_PERIPHERY_CONTRACT: "0xeff187b4190E551FC25a7fA4dFC6cf7fDeF7194f"
+      LENS_PERIPHERY_CONTRACT: "0xeff187b4190E551FC25a7fA4dFC6cf7fDeF7194f",
+      apiLink: "https://api.lens.dev"
     }
   ]
 ]);
@@ -57,15 +53,14 @@ export const defaultConfigs: ProtocolConfig[] = [
     theGraphIpfsUrl: "https://api.thegraph.com/ipfs/api/v0",
     ipfsMetadataUrl: "https://ipfs.infura.io:5001",
     contracts: {
-      // from https://github.com/bosonprotocol/boson-protocol-contracts/commit/2b596b4e9aca6c5b467f5ce324ec1e7dd9fc9c1d
-      protocolDiamond: "0x7Fcf1A9c688E8e0A8d8911D7c6Fb7D1a1D796B80"
+      // from https://github.com/bosonprotocol/boson-protocol-contracts/commit/8be49ba4ff83e80542e4a89bf674b0799f1ee913
+      protocolDiamond: "0x785a225EBAC1b600cA3170C6c7fA3488A203Fc21"
     },
     metaTx: {
       relayerUrl: "https://api.biconomy.io"
     },
     lens: {
-      ...(chainIdToLensContractAddresses.has(80001) &&
-        chainIdToLensContractAddresses.get(80001))
+      ...(chainIdToLensInfo.has(80001) && chainIdToLensInfo.get(80001))
     }
   },
   {
@@ -80,36 +75,35 @@ export const defaultConfigs: ProtocolConfig[] = [
     theGraphIpfsUrl: "https://api.thegraph.com/ipfs/api/v0",
     ipfsMetadataUrl: "https://ipfs.infura.io:5001",
     contracts: {
-      // from https://github.com/bosonprotocol/boson-protocol-contracts/commit/3f5881f66ed2a721a6689103f90a2403d03707b8
-      protocolDiamond: "0x26F5BC2D2B0CA2951f55B12437Bf044Ff0DB6f2d"
+      // from https://github.com/bosonprotocol/boson-protocol-contracts/commit/8be49ba4ff83e80542e4a89bf674b0799f1ee913
+      protocolDiamond: "0x5099CA7839e1580bD0C12FC5FECfA45147886BeB"
     },
     metaTx: {
       relayerUrl: "https://api.biconomy.io"
     },
     lens: {
-      ...(chainIdToLensContractAddresses.has(80001) &&
-        chainIdToLensContractAddresses.get(80001))
+      ...(chainIdToLensInfo.has(80001) && chainIdToLensInfo.get(80001))
     }
   },
   {
     envName: "production",
-    chainId: 80001, // TODO: swap to 137 after protocol deployment
-    nativeCoin: chainIdToInfo.get(80001), // TODO: swap to 137 after protocol deployment
-    getTxExplorerUrl: chainIdToGraphTx.get(80001), // TODO: swap to 137 after protocol deployment
-    subgraphUrl: "https://api.thegraph.com/subgraphs/name/bosonprotocol/mumbai", // TODO: re-deploy subgraphs to correct org & swap for "real" one after protocol deployment
+    chainId: 137,
+    nativeCoin: chainIdToInfo.get(137),
+    getTxExplorerUrl: chainIdToGraphTx.get(137),
+    subgraphUrl:
+      "https://api.thegraph.com/subgraphs/name/bosonprotocol/polygon",
     jsonRpcUrl:
-      "https://polygon-mumbai.infura.io/v3/383117b55d614525b07f03b5979c5f19", // TODO: switch to polygon endpoint after protocol deployment
+      "https://polygon-mainnet.infura.io/v3/383117b55d614525b07f03b5979c5f19",
     theGraphIpfsUrl: "https://api.thegraph.com/ipfs/api/v0",
     ipfsMetadataUrl: "https://ipfs.infura.io:5001",
     contracts: {
-      protocolDiamond: "0x1796B155D4A719d6eBe0496F4914c98a480e668C"
+      protocolDiamond: "0x59A4C19b55193D5a2EAD0065c54af4d516E18Cb5"
     },
     metaTx: {
-      relayerUrl: "https://api.biconomy.io" // TODO: where will we configure the API key? Also for staging config
+      relayerUrl: "https://api.biconomy.io"
     },
     lens: {
-      ...(chainIdToLensContractAddresses.has(80001) &&
-        chainIdToLensContractAddresses.get(80001))
+      ...(chainIdToLensInfo.has(137) && chainIdToLensInfo.get(137))
     }
   },
   {
@@ -129,8 +123,7 @@ export const defaultConfigs: ProtocolConfig[] = [
       relayerUrl: "http://localhost:8888"
     },
     lens: {
-      ...(chainIdToLensContractAddresses.has(31337) &&
-        chainIdToLensContractAddresses.get(31337))
+      ...(chainIdToLensInfo.has(31337) && chainIdToLensInfo.get(31337))
     }
   }
 ];