+++ packages/core-sdk/src/subgraph.ts
@@ -185,6 +185,7 @@ export type BaseMetadataEntity = MetadataInterface & {
    */
   id: Scalars["ID"];
   image: Scalars["String"];
+  licenseUrl: Scalars["String"];
   name: Scalars["String"];
   numberOfCommits: Scalars["BigInt"];
   numberOfRedemptions: Scalars["BigInt"];
@@ -318,6 +319,26 @@ export type BaseMetadataEntity_Filter = {
   image_not_starts_with_nocase?: InputMaybe<Scalars["String"]>;
   image_starts_with?: InputMaybe<Scalars["String"]>;
   image_starts_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl?: InputMaybe<Scalars["String"]>;
+  licenseUrl_contains?: InputMaybe<Scalars["String"]>;
+  licenseUrl_contains_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_ends_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_ends_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_gt?: InputMaybe<Scalars["String"]>;
+  licenseUrl_gte?: InputMaybe<Scalars["String"]>;
+  licenseUrl_in?: InputMaybe<Array<Scalars["String"]>>;
+  licenseUrl_lt?: InputMaybe<Scalars["String"]>;
+  licenseUrl_lte?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_contains?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_contains_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_ends_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_ends_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_in?: InputMaybe<Array<Scalars["String"]>>;
+  licenseUrl_not_starts_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_starts_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_starts_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_starts_with_nocase?: InputMaybe<Scalars["String"]>;
   name?: InputMaybe<Scalars["String"]>;
   name_contains?: InputMaybe<Scalars["String"]>;
   name_contains_nocase?: InputMaybe<Scalars["String"]>;
@@ -458,6 +479,7 @@ export enum BaseMetadataEntity_OrderBy {
   ExternalUrl = "externalUrl",
   Id = "id",
   Image = "image",
+  LicenseUrl = "licenseUrl",
   Name = "name",
   NumberOfCommits = "numberOfCommits",
   NumberOfRedemptions = "numberOfRedemptions",
@@ -1366,6 +1388,7 @@ export enum EventType {
   SellerCreated = "SELLER_CREATED",
   SellerUpdated = "SELLER_UPDATED",
   VoucherCanceled = "VOUCHER_CANCELED",
+  VoucherExpired = "VOUCHER_EXPIRED",
   VoucherExtended = "VOUCHER_EXTENDED",
   VoucherRedeemed = "VOUCHER_REDEEMED",
   VoucherRevoked = "VOUCHER_REVOKED",
@@ -2167,6 +2190,7 @@ export type MetadataInterface = {
    */
   id: Scalars["ID"];
   image: Scalars["String"];
+  licenseUrl: Scalars["String"];
   name: Scalars["String"];
   numberOfCommits: Scalars["BigInt"];
   numberOfRedemptions: Scalars["BigInt"];
@@ -2299,6 +2323,26 @@ export type MetadataInterface_Filter = {
   image_not_starts_with_nocase?: InputMaybe<Scalars["String"]>;
   image_starts_with?: InputMaybe<Scalars["String"]>;
   image_starts_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl?: InputMaybe<Scalars["String"]>;
+  licenseUrl_contains?: InputMaybe<Scalars["String"]>;
+  licenseUrl_contains_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_ends_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_ends_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_gt?: InputMaybe<Scalars["String"]>;
+  licenseUrl_gte?: InputMaybe<Scalars["String"]>;
+  licenseUrl_in?: InputMaybe<Array<Scalars["String"]>>;
+  licenseUrl_lt?: InputMaybe<Scalars["String"]>;
+  licenseUrl_lte?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_contains?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_contains_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_ends_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_ends_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_in?: InputMaybe<Array<Scalars["String"]>>;
+  licenseUrl_not_starts_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_starts_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_starts_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_starts_with_nocase?: InputMaybe<Scalars["String"]>;
   name?: InputMaybe<Scalars["String"]>;
   name_contains?: InputMaybe<Scalars["String"]>;
   name_contains_nocase?: InputMaybe<Scalars["String"]>;
@@ -2439,6 +2483,7 @@ export enum MetadataInterface_OrderBy {
   ExternalUrl = "externalUrl",
   Id = "id",
   Image = "image",
+  LicenseUrl = "licenseUrl",
   Name = "name",
   NumberOfCommits = "numberOfCommits",
   NumberOfRedemptions = "numberOfRedemptions",
@@ -3338,6 +3383,7 @@ export type ProductV1MetadataEntity = MetadataInterface & {
    */
   id: Scalars["ID"];
   image: Scalars["String"];
+  licenseUrl: Scalars["String"];
   name: Scalars["String"];
   numberOfCommits: Scalars["BigInt"];
   numberOfRedemptions: Scalars["BigInt"];
@@ -3510,6 +3556,26 @@ export type ProductV1MetadataEntity_Filter = {
   image_not_starts_with_nocase?: InputMaybe<Scalars["String"]>;
   image_starts_with?: InputMaybe<Scalars["String"]>;
   image_starts_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl?: InputMaybe<Scalars["String"]>;
+  licenseUrl_contains?: InputMaybe<Scalars["String"]>;
+  licenseUrl_contains_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_ends_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_ends_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_gt?: InputMaybe<Scalars["String"]>;
+  licenseUrl_gte?: InputMaybe<Scalars["String"]>;
+  licenseUrl_in?: InputMaybe<Array<Scalars["String"]>>;
+  licenseUrl_lt?: InputMaybe<Scalars["String"]>;
+  licenseUrl_lte?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_contains?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_contains_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_ends_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_ends_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_in?: InputMaybe<Array<Scalars["String"]>>;
+  licenseUrl_not_starts_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_not_starts_with_nocase?: InputMaybe<Scalars["String"]>;
+  licenseUrl_starts_with?: InputMaybe<Scalars["String"]>;
+  licenseUrl_starts_with_nocase?: InputMaybe<Scalars["String"]>;
   name?: InputMaybe<Scalars["String"]>;
   name_contains?: InputMaybe<Scalars["String"]>;
   name_contains_nocase?: InputMaybe<Scalars["String"]>;
@@ -3762,6 +3828,7 @@ export enum ProductV1MetadataEntity_OrderBy {
   ExternalUrl = "externalUrl",
   Id = "id",
   Image = "image",
+  LicenseUrl = "licenseUrl",
   Name = "name",
   NumberOfCommits = "numberOfCommits",
   NumberOfRedemptions = "numberOfRedemptions",
@@ -7487,6 +7554,7 @@ export type GetSellerByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -7502,6 +7570,7 @@ export type GetSellerByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -7928,6 +7997,7 @@ export type GetSellersQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -7943,6 +8013,7 @@ export type GetSellersQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -8671,6 +8742,7 @@ export type GetDisputeResolverByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -8686,6 +8758,7 @@ export type GetDisputeResolverByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -9049,6 +9122,7 @@ export type GetDisputeResolversQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -9064,6 +9138,7 @@ export type GetDisputeResolversQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -9421,6 +9496,7 @@ export type SellerFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
         }
@@ -9436,6 +9512,7 @@ export type SellerFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
           attributes?: Array<{
@@ -9966,6 +10043,7 @@ export type DisputeResolverFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
         }
@@ -9981,6 +10059,7 @@ export type DisputeResolverFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
           attributes?: Array<{
@@ -10682,6 +10761,7 @@ export type GetExchangeTokenByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -10697,6 +10777,7 @@ export type GetExchangeTokenByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -10981,6 +11062,7 @@ export type GetExchangeTokensQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -10996,6 +11078,7 @@ export type GetExchangeTokensQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -11258,6 +11341,7 @@ export type ExchangeTokenFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
         }
@@ -11273,6 +11357,7 @@ export type ExchangeTokenFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
           attributes?: Array<{
@@ -11706,6 +11791,7 @@ export type GetExchangeByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -11721,6 +11807,7 @@ export type GetExchangeByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -12029,6 +12116,7 @@ export type GetExchangesQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -12044,6 +12132,7 @@ export type GetExchangesQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -12342,6 +12431,7 @@ export type ExchangeFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
         }
@@ -12357,6 +12447,7 @@ export type ExchangeFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
           attributes?: Array<{
@@ -12675,6 +12766,7 @@ export type GetBaseMetadataEntityByIdQueryQuery = {
     name: string;
     description: string;
     externalUrl: string;
+    licenseUrl: string;
     schemaUrl: string;
     type: MetadataType;
     image: string;
@@ -12780,6 +12872,7 @@ export type GetBaseMetadataEntityByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -12795,6 +12888,7 @@ export type GetBaseMetadataEntityByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -12993,6 +13087,7 @@ export type GetBaseMetadataEntitiesQueryQuery = {
     name: string;
     description: string;
     externalUrl: string;
+    licenseUrl: string;
     schemaUrl: string;
     type: MetadataType;
     image: string;
@@ -13098,6 +13193,7 @@ export type GetBaseMetadataEntitiesQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -13113,6 +13209,7 @@ export type GetBaseMetadataEntitiesQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -13301,6 +13398,7 @@ export type BaseMetadataEntityFieldsFragment = {
   name: string;
   description: string;
   externalUrl: string;
+  licenseUrl: string;
   schemaUrl: string;
   type: MetadataType;
   image: string;
@@ -13406,6 +13504,7 @@ export type BaseMetadataEntityFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
         }
@@ -13421,6 +13520,7 @@ export type BaseMetadataEntityFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
           attributes?: Array<{
@@ -13604,6 +13704,7 @@ export type BaseBaseMetadataEntityFieldsFragment = {
   name: string;
   description: string;
   externalUrl: string;
+  licenseUrl: string;
   schemaUrl: string;
   type: MetadataType;
   image: string;
@@ -13709,6 +13810,7 @@ export type BaseBaseMetadataEntityFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
         }
@@ -13724,6 +13826,7 @@ export type BaseBaseMetadataEntityFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
           attributes?: Array<{
@@ -13952,6 +14055,7 @@ export type GetProductV1MetadataEntityByIdQueryQuery = {
     name: string;
     description: string;
     externalUrl: string;
+    licenseUrl: string;
     schemaUrl: string;
     type: MetadataType;
     image: string;
@@ -14058,6 +14162,7 @@ export type GetProductV1MetadataEntityByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -14073,6 +14178,7 @@ export type GetProductV1MetadataEntityByIdQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -14399,6 +14505,7 @@ export type GetProductV1MetadataEntitiesQueryQuery = {
     name: string;
     description: string;
     externalUrl: string;
+    licenseUrl: string;
     schemaUrl: string;
     type: MetadataType;
     image: string;
@@ -14505,6 +14612,7 @@ export type GetProductV1MetadataEntitiesQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
           }
@@ -14520,6 +14628,7 @@ export type GetProductV1MetadataEntitiesQueryQuery = {
             name: string;
             description: string;
             externalUrl: string;
+            licenseUrl: string;
             schemaUrl: string;
             type: MetadataType;
             attributes?: Array<{
@@ -14836,6 +14945,7 @@ export type ProductV1MetadataEntityFieldsFragment = {
   name: string;
   description: string;
   externalUrl: string;
+  licenseUrl: string;
   schemaUrl: string;
   type: MetadataType;
   image: string;
@@ -14942,6 +15052,7 @@ export type ProductV1MetadataEntityFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
         }
@@ -14957,6 +15068,7 @@ export type ProductV1MetadataEntityFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
           attributes?: Array<{
@@ -15268,6 +15380,7 @@ export type BaseProductV1MetadataEntityFieldsFragment = {
   name: string;
   description: string;
   externalUrl: string;
+  licenseUrl: string;
   schemaUrl: string;
   type: MetadataType;
   image: string;
@@ -15374,6 +15487,7 @@ export type BaseProductV1MetadataEntityFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
         }
@@ -15389,6 +15503,7 @@ export type BaseProductV1MetadataEntityFieldsFragment = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
           attributes?: Array<{
@@ -16082,6 +16197,7 @@ export type GetOfferByIdQueryQuery = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
         }
@@ -16097,6 +16213,7 @@ export type GetOfferByIdQueryQuery = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
           attributes?: Array<{
@@ -16407,6 +16524,7 @@ export type GetOffersQueryQuery = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
         }
@@ -16422,6 +16540,7 @@ export type GetOffersQueryQuery = {
           name: string;
           description: string;
           externalUrl: string;
+          licenseUrl: string;
           schemaUrl: string;
           type: MetadataType;
           attributes?: Array<{
@@ -16716,6 +16835,7 @@ export type OfferFieldsFragment = {
         name: string;
         description: string;
         externalUrl: string;
+        licenseUrl: string;
         schemaUrl: string;
         type: MetadataType;
       }
@@ -16731,6 +16851,7 @@ export type OfferFieldsFragment = {
         name: string;
         description: string;
         externalUrl: string;
+        licenseUrl: string;
         schemaUrl: string;
         type: MetadataType;
         attributes?: Array<{
@@ -16974,6 +17095,7 @@ export type BaseOfferFieldsFragment = {
         name: string;
         description: string;
         externalUrl: string;
+        licenseUrl: string;
         schemaUrl: string;
         type: MetadataType;
       }
@@ -16989,6 +17111,7 @@ export type BaseOfferFieldsFragment = {
         name: string;
         description: string;
         externalUrl: string;
+        licenseUrl: string;
         schemaUrl: string;
         type: MetadataType;
         attributes?: Array<{
@@ -17434,6 +17557,7 @@ export const BaseOfferFieldsFragmentDoc = gql`
       name
       description
       externalUrl
+      licenseUrl
       schemaUrl
       type
       ... on ProductV1MetadataEntity {
@@ -17728,6 +17852,7 @@ export const BaseBaseMetadataEntityFieldsFragmentDoc = gql`
     name
     description
     externalUrl
+    licenseUrl
     schemaUrl
     type
     image
@@ -17769,6 +17894,7 @@ export const BaseProductV1MetadataEntityFieldsFragmentDoc = gql`
     name
     description
     externalUrl
+    licenseUrl
     schemaUrl
     type
     image