+++ e2e/meta-tx-gateway/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.4](https://github.com/bosonprotocol/core-components/compare/meta-tx-gateway@1.1.3...meta-tx-gateway@1.1.4) (2022-10-13)
+
+**Note:** Version bump only for package meta-tx-gateway
+
+
+
+
+
+## [1.1.3](https://github.com/bosonprotocol/core-components/compare/meta-tx-gateway@1.1.2...meta-tx-gateway@1.1.3) (2022-10-13)
+
+**Note:** Version bump only for package meta-tx-gateway
+
+
+
+
+
+## [1.1.2](https://github.com/bosonprotocol/core-components/compare/meta-tx-gateway@1.1.1...meta-tx-gateway@1.1.2) (2022-10-12)
+
+**Note:** Version bump only for package meta-tx-gateway
+
+
+
+
+
 ## [1.1.1](https://github.com/bosonprotocol/core-components/compare/meta-tx-gateway@1.1.0...meta-tx-gateway@1.1.1) (2022-10-05)
 
 **Note:** Version bump only for package meta-tx-gateway