+++ packages/ethers-sdk/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.2](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/ethers-sdk@1.9.1...@bosonprotocol/ethers-sdk@1.9.2) (2022-10-13)
+
+**Note:** Version bump only for package @bosonprotocol/ethers-sdk
+
+
+
+
+
+## [1.9.1](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/ethers-sdk@1.9.0...@bosonprotocol/ethers-sdk@1.9.1) (2022-10-13)
+
+**Note:** Version bump only for package @bosonprotocol/ethers-sdk
+
+
+
+
+
+# [1.9.0](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/ethers-sdk@1.8.0...@bosonprotocol/ethers-sdk@1.9.0) (2022-10-12)
+
+
+### Features
+
+* contract uri and royalty support in subgraph ([#303](https://github.com/bosonprotocol/core-components/issues/303)) ([edc44c1](https://github.com/bosonprotocol/core-components/commit/edc44c1977190635becfb8110d39396545feb5df))
+* update to latest commit of contracts ([#297](https://github.com/bosonprotocol/core-components/issues/297)) ([0efcc9c](https://github.com/bosonprotocol/core-components/commit/0efcc9ca05dd85bffbd48fa927f69667c605f708))
+
+
+
+
+
 # [1.8.0](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/ethers-sdk@1.7.1...@bosonprotocol/ethers-sdk@1.8.0) (2022-10-05)